refactor: introduce application layer in Compiler
