+++ src/phel/core.phel
@@ -29,7 +29,7 @@
   (:use Phel\Lang\Variable)
   (:use Phel\Compiler\CompilerFacade)
   (:use Phel\Compiler\Infrastructure\CompileOptions)
-  (:use Phel\Compiler\Infrastructure\Munge)
+  (:use Phel\Compiler\Application\Munge)
   (:use Phel\Printer\Printer)
   (:use Countable)
   (:use Exception)
