+++ src/phel/test.phel
@@ -1,7 +1,7 @@
 (ns phel\test
   (:use Phel\Lang\Symbol)
   (:use Phel\Lang\Registry)
-  (:use Phel\Compiler\Infrastructure\Munge)
+  (:use Phel\Compiler\Application\Munge)
   (:use Phel\Command\CommandFacade))
 
 # ------
