+++ src/php/Command/CommandFactory.php
@@ -19,7 +19,7 @@ use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
 use Phel\Command\Domain\Shared\Exceptions\Extractor\FilePositionExtractor;
 use Phel\Command\Domain\Shared\Exceptions\Extractor\SourceMapExtractor;
 use Phel\Command\Domain\Shared\Exceptions\TextExceptionPrinter;
-use Phel\Compiler\Infrastructure\Munge;
+use Phel\Compiler\Application\Munge;
 use Phel\Printer\Printer;
 use Phel\Run\Domain\Repl\ColorStyle;
