+++ src/php/Compiler/Application/CodeCompiler.php
@@ -2,12 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Phel\Compiler\Domain\Compiler;
+namespace Phel\Compiler\Application;
 
 use Phel\Compiler\Domain\Analyzer\AnalyzerInterface;
 use Phel\Compiler\Domain\Analyzer\Ast\AbstractNode;
 use Phel\Compiler\Domain\Analyzer\Environment\NodeEnvironment;
 use Phel\Compiler\Domain\Analyzer\Exceptions\AnalyzerException;
+use Phel\Compiler\Domain\Compiler\CodeCompilerInterface;
+use Phel\Compiler\Domain\Emitter\EmitterResult;
+use Phel\Compiler\Domain\Emitter\FileEmitterInterface;
 use Phel\Compiler\Domain\Emitter\StatementEmitterInterface;
 use Phel\Compiler\Domain\Evaluator\EvaluatorInterface;
 use Phel\Compiler\Domain\Evaluator\Exceptions\CompiledCodeIsMalformedException;
@@ -16,7 +19,6 @@ use Phel\Compiler\Domain\Exceptions\CompilerException;
 use Phel\Compiler\Domain\Lexer\Exceptions\LexerValueException;
 use Phel\Compiler\Domain\Lexer\LexerInterface;
 use Phel\Compiler\Domain\Parser\Exceptions\AbstractParserException;
-use Phel\Compiler\Domain\Parser\Exceptions\UnfinishedParserException;
 use Phel\Compiler\Domain\Parser\ParserInterface;
 use Phel\Compiler\Domain\Parser\ParserNode\NodeInterface;
 use Phel\Compiler\Domain\Parser\ParserNode\TriviaNodeInterface;
@@ -26,60 +28,64 @@ use Phel\Compiler\Domain\Reader\ReaderInterface;
 use Phel\Compiler\Infrastructure\CompileOptions;
 use Phel\Lang\TypeInterface;
 
-final readonly class EvalCompiler implements EvalCompilerInterface
+final readonly class CodeCompiler implements CodeCompilerInterface
 {
     public function __construct(
         private LexerInterface $lexer,
         private ParserInterface $parser,
         private ReaderInterface $reader,
         private AnalyzerInterface $analyzer,
-        private StatementEmitterInterface $emitter,
+        private StatementEmitterInterface $statementEmitter,
+        private FileEmitterInterface $fileEmitter,
         private EvaluatorInterface $evaluator,
     ) {
     }
 
     /**
-     * Evaluates a provided Phel code.
-     *
-     * @throws CompiledCodeIsMalformedException
      * @throws CompilerException
+     * @throws CompiledCodeIsMalformedException
      * @throws FileException
      * @throws LexerValueException
-     * @throws UnfinishedParserException
-     *
-     * @return mixed The result of the executed code
      */
-    public function evalString(string $phelCode, CompileOptions $compileOptions): mixed
+    public function compileString(string $phelCode, CompileOptions $compileOptions): EmitterResult
     {
-        $tokenStream = $this->lexer->lexString($phelCode, $compileOptions->getSource(), $compileOptions->getStartingLine());
+        $tokenStream = $this->lexer->lexString(
+            $phelCode,
+            $compileOptions->getSource(),
+            $compileOptions->getStartingLine(),
+        );
 
-        $result = null;
+        $this->fileEmitter->startFile($compileOptions->getSource());
         while (true) {
             try {
                 $parseTree = $this->parser->parseNext($tokenStream);
-
+                // If we reached the end exit this loop
                 if (!$parseTree instanceof NodeInterface) {
-                    return $result;
+                    break;
                 }
 
                 if (!$parseTree instanceof TriviaNodeInterface) {
                     $readerResult = $this->reader->read($parseTree);
                     $node = $this->analyze($readerResult);
-
-                    $result = $this->evalNode($node, $compileOptions);
+                    // We need to evaluate every statement because we may need it for macros.
+                    $this->emitNode($node, $compileOptions);
                 }
-            } catch (UnfinishedParserException $e) {
-                throw $e;
             } catch (AbstractParserException|ReaderException $e) {
                 throw new CompilerException($e, $e->getCodeSnippet());
             }
         }
+
+        return $this->fileEmitter->endFile($compileOptions->isSourceMapsEnabled());
     }
 
-    public function evalForm(float|bool|int|string|TypeInterface|null $form, CompileOptions $compileOptions): mixed
-    {
-        $node = $this->analyzer->analyze($form, NodeEnvironment::empty()->withReturnContext());
-        return $this->evalNode($node, $compileOptions);
+    public function compileForm(
+        float|bool|int|string|TypeInterface|null $form,
+        CompileOptions $compileOptions,
+    ): EmitterResult {
+        $this->fileEmitter->startFile($compileOptions->getSource());
+        $node = $this->analyzer->analyze($form, NodeEnvironment::empty());
+        $this->emitNode($node, $compileOptions);
+        return $this->fileEmitter->endFile($compileOptions->isSourceMapsEnabled());
     }
 
     /**
@@ -90,7 +96,7 @@ final readonly class EvalCompiler implements EvalCompilerInterface
         try {
             return $this->analyzer->analyze(
                 $readerResult->getAst(),
-                NodeEnvironment::empty()->withReturnContext(),
+                NodeEnvironment::empty(),
             );
         } catch (AnalyzerException $analyzerException) {
             throw new CompilerException($analyzerException, $readerResult->getCodeSnippet());
@@ -100,13 +106,16 @@ final readonly class EvalCompiler implements EvalCompilerInterface
     /**
      * @throws CompiledCodeIsMalformedException
      * @throws FileException
-     *
-     * @return mixed The result of the executed code
      */
-    private function evalNode(AbstractNode $node, CompileOptions $compileOptions): mixed
+    private function emitNode(AbstractNode $node, CompileOptions $compileOptions): void
     {
-        $code = $this->emitter->emitNode($node, $compileOptions->isSourceMapsEnabled())->getCodeWithSourceMap();
+        $this->fileEmitter->emitNode($node);
 
-        return $this->evaluator->eval($code);
+        $code = $this->statementEmitter->emitNode(
+            $node,
+            $compileOptions->isSourceMapsEnabled(),
+        )->getCodeWithSourceMap();
+
+        $this->evaluator->eval($code);
     }
 }
