+++ src/php/Compiler/Application/EvalCompiler.php
@@ -2,14 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Phel\Compiler\Domain\Compiler;
+namespace Phel\Compiler\Application;
 
 use Phel\Compiler\Domain\Analyzer\AnalyzerInterface;
 use Phel\Compiler\Domain\Analyzer\Ast\AbstractNode;
 use Phel\Compiler\Domain\Analyzer\Environment\NodeEnvironment;
 use Phel\Compiler\Domain\Analyzer\Exceptions\AnalyzerException;
-use Phel\Compiler\Domain\Emitter\EmitterResult;
-use Phel\Compiler\Domain\Emitter\FileEmitterInterface;
+use Phel\Compiler\Domain\Compiler\EvalCompilerInterface;
 use Phel\Compiler\Domain\Emitter\StatementEmitterInterface;
 use Phel\Compiler\Domain\Evaluator\EvaluatorInterface;
 use Phel\Compiler\Domain\Evaluator\Exceptions\CompiledCodeIsMalformedException;
@@ -18,6 +17,7 @@ use Phel\Compiler\Domain\Exceptions\CompilerException;
 use Phel\Compiler\Domain\Lexer\Exceptions\LexerValueException;
 use Phel\Compiler\Domain\Lexer\LexerInterface;
 use Phel\Compiler\Domain\Parser\Exceptions\AbstractParserException;
+use Phel\Compiler\Domain\Parser\Exceptions\UnfinishedParserException;
 use Phel\Compiler\Domain\Parser\ParserInterface;
 use Phel\Compiler\Domain\Parser\ParserNode\NodeInterface;
 use Phel\Compiler\Domain\Parser\ParserNode\TriviaNodeInterface;
@@ -27,58 +27,60 @@ use Phel\Compiler\Domain\Reader\ReaderInterface;
 use Phel\Compiler\Infrastructure\CompileOptions;
 use Phel\Lang\TypeInterface;
 
-final readonly class CodeCompiler implements CodeCompilerInterface
+final readonly class EvalCompiler implements EvalCompilerInterface
 {
     public function __construct(
         private LexerInterface $lexer,
         private ParserInterface $parser,
         private ReaderInterface $reader,
         private AnalyzerInterface $analyzer,
-        private StatementEmitterInterface $statementEmitter,
-        private FileEmitterInterface $fileEmitter,
+        private StatementEmitterInterface $emitter,
         private EvaluatorInterface $evaluator,
     ) {
     }
 
     /**
-     * @throws CompilerException
+     * Evaluates a provided Phel code.
+     *
      * @throws CompiledCodeIsMalformedException
+     * @throws CompilerException
      * @throws FileException
      * @throws LexerValueException
+     * @throws UnfinishedParserException
+     *
+     * @return mixed The result of the executed code
      */
-    public function compileString(string $phelCode, CompileOptions $compileOptions): EmitterResult
+    public function evalString(string $phelCode, CompileOptions $compileOptions): mixed
     {
         $tokenStream = $this->lexer->lexString($phelCode, $compileOptions->getSource(), $compileOptions->getStartingLine());
 
-        $this->fileEmitter->startFile($compileOptions->getSource());
+        $result = null;
         while (true) {
             try {
                 $parseTree = $this->parser->parseNext($tokenStream);
-                // If we reached the end exit this loop
+
                 if (!$parseTree instanceof NodeInterface) {
-                    break;
+                    return $result;
                 }
 
                 if (!$parseTree instanceof TriviaNodeInterface) {
                     $readerResult = $this->reader->read($parseTree);
                     $node = $this->analyze($readerResult);
-                    // We need to evaluate every statement because we may need it for macros.
-                    $this->emitNode($node, $compileOptions);
+
+                    $result = $this->evalNode($node, $compileOptions);
                 }
+            } catch (UnfinishedParserException $e) {
+                throw $e;
             } catch (AbstractParserException|ReaderException $e) {
                 throw new CompilerException($e, $e->getCodeSnippet());
             }
         }
-
-        return $this->fileEmitter->endFile($compileOptions->isSourceMapsEnabled());
     }
 
-    public function compileForm(float|bool|int|string|TypeInterface|null $form, CompileOptions $compileOptions): EmitterResult
+    public function evalForm(float|bool|int|string|TypeInterface|null $form, CompileOptions $compileOptions): mixed
     {
-        $this->fileEmitter->startFile($compileOptions->getSource());
-        $node = $this->analyzer->analyze($form, NodeEnvironment::empty());
-        $this->emitNode($node, $compileOptions);
-        return $this->fileEmitter->endFile($compileOptions->isSourceMapsEnabled());
+        $node = $this->analyzer->analyze($form, NodeEnvironment::empty()->withReturnContext());
+        return $this->evalNode($node, $compileOptions);
     }
 
     /**
@@ -89,7 +91,7 @@ final readonly class CodeCompiler implements CodeCompilerInterface
         try {
             return $this->analyzer->analyze(
                 $readerResult->getAst(),
-                NodeEnvironment::empty(),
+                NodeEnvironment::empty()->withReturnContext(),
             );
         } catch (AnalyzerException $analyzerException) {
             throw new CompilerException($analyzerException, $readerResult->getCodeSnippet());
@@ -99,12 +101,13 @@ final readonly class CodeCompiler implements CodeCompilerInterface
     /**
      * @throws CompiledCodeIsMalformedException
      * @throws FileException
+     *
+     * @return mixed The result of the executed code
      */
-    private function emitNode(AbstractNode $node, CompileOptions $compileOptions): void
+    private function evalNode(AbstractNode $node, CompileOptions $compileOptions): mixed
     {
-        $this->fileEmitter->emitNode($node);
+        $code = $this->emitter->emitNode($node, $compileOptions->isSourceMapsEnabled())->getCodeWithSourceMap();
 
-        $code = $this->statementEmitter->emitNode($node, $compileOptions->isSourceMapsEnabled())->getCodeWithSourceMap();
-        $this->evaluator->eval($code);
+        return $this->evaluator->eval($code);
     }
 }
