+++ src/php/Compiler/CompilerFactory.php
@@ -5,12 +5,16 @@ declare(strict_types=1);
 namespace Phel\Compiler;
 
 use Gacela\Framework\AbstractFactory;
-use Phel\Compiler\Domain\Analyzer\Analyzer;
+use Phel\Compiler\Application\Analyzer;
+use Phel\Compiler\Application\CodeCompiler;
+use Phel\Compiler\Application\EvalCompiler;
+use Phel\Compiler\Application\Lexer;
+use Phel\Compiler\Application\Munge;
+use Phel\Compiler\Application\Parser;
+use Phel\Compiler\Application\Reader;
 use Phel\Compiler\Domain\Analyzer\AnalyzerInterface;
 use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironmentInterface;
-use Phel\Compiler\Domain\Compiler\CodeCompiler;
 use Phel\Compiler\Domain\Compiler\CodeCompilerInterface;
-use Phel\Compiler\Domain\Compiler\EvalCompiler;
 use Phel\Compiler\Domain\Compiler\EvalCompilerInterface;
 use Phel\Compiler\Domain\Emitter\FileEmitter;
 use Phel\Compiler\Domain\Emitter\FileEmitterInterface;
@@ -25,18 +29,14 @@ use Phel\Compiler\Domain\Emitter\StatementEmitter;
 use Phel\Compiler\Domain\Emitter\StatementEmitterInterface;
 use Phel\Compiler\Domain\Evaluator\EvaluatorInterface;
 use Phel\Compiler\Domain\Evaluator\RequireEvaluator;
-use Phel\Compiler\Domain\Lexer\Lexer;
 use Phel\Compiler\Domain\Lexer\LexerInterface;
 use Phel\Compiler\Domain\Parser\ExpressionParserFactory;
-use Phel\Compiler\Domain\Parser\Parser;
 use Phel\Compiler\Domain\Parser\ParserInterface;
 use Phel\Compiler\Domain\Reader\ExpressionReaderFactory;
 use Phel\Compiler\Domain\Reader\QuasiquoteTransformer;
-use Phel\Compiler\Domain\Reader\Reader;
 use Phel\Compiler\Domain\Reader\ReaderInterface;
 use Phel\Compiler\Infrastructure\CompileOptions;
 use Phel\Compiler\Infrastructure\GlobalEnvironmentSingleton;
-use Phel\Compiler\Infrastructure\Munge;
 use Phel\Filesystem\FilesystemFacadeInterface;
 use Phel\Printer\Printer;
