+++ src/php/Compiler/Domain/Analyzer/Ast/PhpClassNameNode.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Domain\Analyzer\Ast;
 
+use Phel\Compiler\Application\Munge;
 use Phel\Compiler\Domain\Analyzer\Environment\NodeEnvironmentInterface;
-use Phel\Compiler\Infrastructure\Munge;
 use Phel\Lang\SourceLocation;
 use Phel\Lang\Symbol;
 use ReflectionClass;
